# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Couture POS API",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "ShiftHistoryEntry": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Shift"
          },
          {
            "type": "object",
            "properties": {
              "staffName": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "terminalName": {
                "type": [
                  "string",
                  "null"
                ]
              }
            },
            "required": [
              "staffName",
              "terminalName"
            ]
          }
        ]
      },
      "Shift": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "staffId": {
            "type": "string",
            "format": "uuid"
          },
          "terminalId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "startingCash": {
            "type": "string"
          },
          "openedAt": {
            "type": "string"
          },
          "countedCash": {
            "type": [
              "string",
              "null"
            ]
          },
          "variance": {
            "type": [
              "string",
              "null"
            ]
          },
          "closedAt": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "staffId",
          "terminalId",
          "startingCash",
          "openedAt",
          "countedCash",
          "variance",
          "closedAt"
        ]
      },
...
      "OpenShiftRequest": {
        "type": "object",
        "properties": {
          "startingCash": {
            "type": "string",
            "pattern": "^\\d+\\.\\d{2}$"
          },
          "terminalId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "XReport": {
        "type": "object",
        "properties": {
          "shiftId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedCash": {
            "type": "string"
          },
          "cashSalesTotal": {
            "type": "string"
          },
          "cardSalesTotal": {
            "type": "string"
          },
          "checkSalesTotal": {
            "type": "string"
          },
          "refundsTotal": {
            "type": "string"
          },
          "saleCount": {
            "type": "integer"
          }
        },
        "required": [
          "shiftId",
          "expectedCash",
          "cashSalesTotal",
          "cardSalesTotal",
          "checkSalesTotal",
          "refundsTotal",
          "saleCount"
        ]
      },
      "ZReport": {
        "allOf": [
          {
            "$ref": "#/components/schemas/XReport"
          },
          {
            "type": "object",
            "properties": {
              "countedCash": {
                "type": "string"
              },
              "variance": {
                "type": "string"
              },
              "closedAt": {
                "type": "string"
              }
            },
            "required": [
              "countedCash",
              "variance",
              "closedAt"
            ]
          }
        ]
      },
      "CloseShiftRequest": {
        "type": "object",
        "properties": {
          "countedCash": {
            "type": "string",
            "pattern": "^\\d+\\.\\d{2}$"
          }
        },
        "required": [
          "countedCash"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/shifts": {
      "get": {
        "description": "Shift history for the tenant, newest first, with staff and counter names resolved.",
        "responses": {
          "200": {
            "description": "Shift history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShiftHistoryEntry"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Open a shift with a starting cash count on a named counter (D-14, 0034). One counter holds at most one open shift.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenShiftRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shift opened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shift"
                }
              }
            }
          },
          "404": {
            "description": "Counter not found"
          },
          "409": {
            "description": "That counter already has an open shift, or is turned off"
          }
        }
      }
    },
...
    "/shifts/{shiftId}/x-report": {
      "get": {
        "description": "Live, non-resetting shift snapshot (D-15, CASH-02).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "shiftId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "X report",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/XReport"
                }
              }
            }
          },
          "404": {
            "description": "Shift not found"
          }
        }
      }
    },
    "/shifts/{shiftId}/close": {
      "post": {
        "description": "Close a shift (Z report) — counted cash, variance, locks the shift (D-15/D-16).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "shiftId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseShiftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Shift closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZReport"
                }
              }
            }
          },
          "404": {
            "description": "Shift not found"
          },
          "409": {
            "description": "Shift already closed"
          }
        }
      }
    },
...
  },
...
}